+++ provider/eastmoney/eastmoney_test.go
@@ -12,9 +12,11 @@ func TestMarketType_String(t *testing.T) {
 	var m1 MarketType = 0
 	var m2 MarketType = 1
 	var m3 MarketType = 3
+	var m4 MarketType = 105
 	require.Equal(t, "SZ", m1.String())
 	require.Equal(t, "SH", m2.String())
 	require.Equal(t, "unknown 3", m3.String())
+	require.Equal(t, "$", m4.String())
 }
 
 func TestGetOriginQuoteHistory(t *testing.T) {