+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### v0.2.8
 
 1. 更新`quote-history`命令描述
+2. `quote-history` 支持港股
 
 ### v0.2.7